smart-document-classifier/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
spacy>=3.6.0
PyPDF2>=3.0.0
python-docx>=0.8.11
joblib>=1.3.0
